Refactoring form tests
